Add a filter for proper cache stats after warm-up
The filter will pass a modified AccessInfo to the StatsCounters. The
hit counts are amended. The first time a byte which was already in the
cache at start of the evaluation period (after warm-up) is hit, this
is changed to be a miss.
This ensures that the hit counts match up with the unique bytes count,
and subsequently the hit rates match up with the theoretical best hit
rates.

Tests on a small access trace (two months of PAGS, first month warm-up,
62939 accesses during evaluation period) have shown a 5 % overhead.
This is expected to go down on larger traces, where the filtered data
shrinks over time and the filter is eventually completely disabled when
there is no data left